Refactor `useArgs`/`useGlobals` out
And add a `useStoriesAndArgs`
